+++ tests/functional/test_builtin.py
@@ -105,7 +105,8 @@ class BuiltinAppTest(unittest.TestCase):
                 if resp.status_code == 200:
                     if resp.json["status"] in JOB_STATUS_CATEGORIES[STATUS_CATEGORY_RUNNING]:
                         continue
-                    assert resp.json["status"] == STATUS_SUCCEEDED
+                    assert resp.json["status"] == STATUS_SUCCEEDED, \
+                        "Process execution failed. Response body:\n{}".format(resp.json)
                     resp = self.app.get("{}/result".format(job_url), headers=self.json_headers)
                     assert resp.status_code == 200
                     assert resp.json["outputs"][0]["id"] == "output"